+++ packages/perseus/src/server-item-renderer.tsx
@@ -19,7 +19,6 @@ import HintsRenderer from "./hints-renderer";
 import LoadingContext from "./loading-context";
 import {ApiOptions} from "./perseus-api";
 import Renderer from "./renderer";
-import {scorePerseusItem} from "./renderer-util";
 import Util from "./util";
 
 import type {PerseusItem, ShowSolutions} from "./perseus-types";
@@ -137,13 +136,11 @@ export class ServerItemRenderer
     }
 
     componentDidUpdate(prevProps: Props, prevState: State) {
-        if (this.props.apiOptions.answerableCallback) {
+        const answerableCallback = this.props.apiOptions.answerableCallback;
+        if (answerableCallback != null) {
             const isAnswerable =
                 this.questionRenderer.emptyWidgets().length === 0;
-            const {answerableCallback} = this.props.apiOptions;
-            if (answerableCallback) {
-                answerableCallback(isAnswerable);
-            }
+            answerableCallback(isAnswerable);
         }
 
         if (!this._fullyRendered) {
@@ -347,41 +344,6 @@ export class ServerItemRenderer
      */
     getUserInput(): UserInputMap {
         return this.questionRenderer.getUserInputMap();
-    }
-
-    /**
-     * Grades the item.
-     *
-     * @deprecated use scorePerseusItem
-     */
-    scoreInput(): KEScore {
-        const guess = this.getUserInput();
-        const score = scorePerseusItem(
-            this.props.item.question,
-            guess,
-            this.context.strings,
-            this.context.locale,
-        );
-
-        // Continue to include an empty guess for the now defunct answer area.
-        // TODO(alex): Check whether we rely on the format here for
-        //             analyzing ProblemLogs. If not, remove this layer.
-        const maxCompatGuess = [this.questionRenderer.getUserInput(), []];
-
-        const keScore = Util.keScoreFromPerseusScore(
-            score,
-            maxCompatGuess,
-            this.questionRenderer.getSerializedState(),
-        );
-
-        const emptyQuestionAreaWidgets = this.questionRenderer.emptyWidgets();
-
-        this.setState({
-            questionCompleted: keScore.correct,
-            questionHighlightedWidgets: emptyQuestionAreaWidgets,
-        });
-
-        return keScore;
     }
 
     /**