+++ packages/perseus/src/types.ts
@@ -1,5 +1,4 @@
 import type {ILogger} from "./logging/log";
-import type {Item} from "./multi-items/item-types";
 import type {
     Hint,
     PerseusAnswerArea,
@@ -90,7 +89,6 @@ export interface Widget {
     getUserInput?: () => UserInputArray | UserInput | undefined;
 
     showRationalesForCurrentlySelectedChoices?: (options?: any) => void;
-    examples?: () => ReadonlyArray<string>;
     getPromptJSON?: () => WidgetPromptJSON;
 }
 
@@ -139,8 +137,6 @@ export type ChangeHandler = (
         question?: any;
         answerArea?: PerseusAnswerArea | null;
         itemDataVersion?: Version;
-        // used in MutirenderEditor
-        item?: Item;
         editorMode?: EditorMode;
         jsonMode?: boolean;
         // perseus-all-package/widgets/unit.jsx
@@ -579,6 +575,8 @@ export type WidgetTransform = (
     strings: PerseusStrings,
     problemNumber?: number,
 ) => any;
+
+export type ValidationResult = Extract<PerseusScore, {type: "invalid"}> | null;
 
 export type WidgetScorerFunction = (
     // The user data needed to score